Update the signatures of Collections and CommentGrid
The CommentGrid doesn't always show images (depending on the design).
We update its signature, with the new argument shouldShowGridImages, to make that configurable.

The Collection is now taking the frontId to decide whether or not the CommentGrid should
show the images. The mapping is done by frontIdShouldShowCommentGridImages (name
can be changed later)